requirements.txt
yt-dlp>=2025.1.15
Pillow>=10.0.0
